chore: productise part 11
Address fixme in productise part 9
Change function prototypes for
 - sleep_millis
 - sleep_micros
to use long instead of int64_t

This fixes issues on 32 bit ARM platform.

+++ src/touch_screen.h
@@ -4,7 +4,7 @@
 #include <tslib.h>
 
 typedef struct {
-    unsigned  sleeptime;
+    long sleeptime;
     void (*up)(int x , int y);
     void (*down)(int x , int y);
     void (*motion)(int x , int y);
